requirements.txt
google-adk>=1.0.0
google-genai>=0.7.0
google-cloud-firestore>=2.13.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0
resend>=0.8.0
google-auth>=2.0.0